Refactor RegisterForm into a modular functional component
- Move RegisterForm into its own folder with styles.js, hooks.js
- Convert class component to functional component
- Extract formSubmitted logic into useRegistrationFormSubmit hook
- Add PURPLE_TINT colour token to COLOURS.js
- Add unit tests covering form render, session loading, submit success/failure, and banner content
