+++ src/react/src/components/RegisterForm/hooks.js
@@ -1,0 +1,21 @@
+import { useState, useEffect, useRef } from 'react';
+
+const useRegistrationFormSubmit = (fetching, error, clearForm) => {
+    const [formSubmitted, setFormSubmitted] = useState(false);
+    const prevFetchingRef = useRef(fetching);
+
+    useEffect(() => {
+        const wasFetching = prevFetchingRef.current;
+        prevFetchingRef.current = fetching;
+
+        if (!error && !fetching && wasFetching) {
+            setFormSubmitted(true);
+        }
+    });
+
+    useEffect(() => () => clearForm(), [clearForm]);
+
+    return formSubmitted;
+};
+
+export default useRegistrationFormSubmit;
